this commit represents a large amount of work.  this work includes: -the final unit test stubs for CampaignController -combining two test files into one tests file that were testing the same handler, but the tests files have two different names. -alignment of handler unit test namspaces with their handlers they test -deletion of duped unit tests between two different classes testing the same thing -moving some quries/commands/handlers from sync to async -renaming class files to match their .cs file on disk and vice-versa. -renaming unit test files to after what they're testing -removed any unused using statements and general code layout cleanup
