+++ AllReadyApp/Web-App/AllReady.UnitTest/Areas/Admin/Features/Campaigns/DeleteCampaignCommandHandlerTests.cs
@@ -1,26 +1,29 @@
-﻿using AllReady.Areas.Admin.Features.Campaigns;
-using AllReady.Models;
-using Microsoft.Extensions.DependencyInjection;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using AllReady.Areas.Admin.Features.Campaigns;
+using AllReady.Models;
+using AllReady.UnitTest.Features.Campaigns;
+using Microsoft.Extensions.DependencyInjection;
 using Xunit;
 
-namespace AllReady.UnitTest.Campaigns
+namespace AllReady.UnitTest.Areas.Admin.Features.Campaigns
 {
     public class DeleteCampaignCommandHandlerTests : InMemoryContextTest
     {
         protected override void LoadTestData()
         {
             var context = ServiceProvider.GetService<AllReadyContext>();
-            Organization htb = new Organization()
+
+            var htb = new Organization
             {
                 Name = "Humanitarian Toolbox",
                 LogoUrl = "http://www.htbox.org/upload/home/ht-hero.png",
                 WebUrl = "http://www.htbox.org",
                 Campaigns = new List<Campaign>()
             };
-            Campaign firePrev = new Campaign()
+
+            var firePrev = new Campaign
             {
                 Id = 1,
                 Name = "Neighborhood Fire Prevention Days",
@@ -35,9 +38,9 @@ namespace AllReady.UnitTest.Campaigns
         public async Task ExistingCampaign()
         {
             var context = ServiceProvider.GetService<AllReadyContext>();
-            var query = new DeleteCampaignCommand { CampaignId = 1 };
+            var command = new DeleteCampaignCommand { CampaignId = 1 };
             var handler = new DeleteCampaignCommandHandler(context);
-            await handler.Handle(query);
+            await handler.Handle(command);
 
             var data = context.Campaigns.Count(_ => _.Id == 1);
             Assert.Equal(0, data);
@@ -47,9 +50,29 @@ namespace AllReady.UnitTest.Campaigns
         public async Task CampaignDoesNotExist()
         {
             var context = ServiceProvider.GetService<AllReadyContext>();
-            var query = new DeleteCampaignCommand { CampaignId = 0 };
+            var command = new DeleteCampaignCommand { CampaignId = 0 };
             var handler = new DeleteCampaignCommandHandler(context);
-            await handler.Handle(query);
+            await handler.Handle(command);
+        }
+
+        [Fact]
+        public async Task CampaignIsDeleted()
+        {
+            const int campaignId = 1;
+
+            var sut = new DeleteCampaignCommandHandler(Context);
+            await sut.Handle(new DeleteCampaignCommand { CampaignId = campaignId });
+            Assert.False(Context.Activities.Any(t => t.Id == campaignId));
+        }
+
+        [Fact]
+        public async Task NonExistantTaskDoesNotCauseException()
+        {
+            const int campaignId = 1;
+
+            var sut = new DeleteCampaignCommandHandler(Context);
+            await sut.Handle(new DeleteCampaignCommand { CampaignId = 666 });
+            Assert.False(Context.Activities.Any(t => t.Id == campaignId));
         }
     }
 }
