+++ AllReadyApp/Web-App/AllReady.UnitTest/Areas/Admin/Features/Campaigns/LockUnlockCampaignCommandHandlerTests.cs
@@ -1,6 +1,7 @@
 ﻿using AllReady.Areas.Admin.Features.Campaigns;
 using System.Linq;
 using System.Threading.Tasks;
+using AllReady.UnitTest.Features.Campaigns;
 using Xunit;
 
 namespace AllReady.UnitTest.Areas.Admin.Features.Campaigns
