+++ AllReadyApp/Web-App/AllReady.UnitTest/Areas/Admin/Features/Skills/SkillDeleteCommandHandlerAsyncTests.cs
@@ -1,6 +1,7 @@
 ﻿using AllReady.Areas.Admin.Features.Skills;
 using System.Linq;
 using System.Threading.Tasks;
+using AllReady.UnitTest.Features.Campaigns;
 using Xunit;
 
 namespace AllReady.UnitTest.Areas.Admin.Features.Skills
