+++ AllReadyApp/Web-App/AllReady.UnitTest/Areas/Admin/Features/Skills/SkillDeleteQueryHandlerAsyncTests.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using AllReady.UnitTest.Features.Campaigns;
 using Xunit;
 
 namespace AllReady.UnitTest.Areas.Admin.Features.Skills
