+++ AllReadyApp/Web-App/AllReady.UnitTest/Areas/Admin/Features/Skills/SkillListQueryHandlerAsyncTests.cs
@@ -2,6 +2,7 @@
 using AllReady.Models;
 using System.Linq;
 using System.Threading.Tasks;
+using AllReady.UnitTest.Features.Campaigns;
 using Xunit;
 
 namespace AllReady.UnitTest.Areas.Admin.Features.Skills
