--- AllReadyApp/Web-App/AllReady.UnitTest/Campaigns/EditingCampaign.cs
@@ -1,17 +1,0 @@
-using AllReady.Areas.Admin.Features.Campaigns;
-using AllReady.Areas.Admin.Models;
-using Xunit;
-
-namespace AllReady.UnitTest.Campaigns
-{
-    public class EditingCampaign : InMemoryContextTest
-    {
-        [Fact]
-        public void ModelIsCreated()
-        {
-            var sut = new EditCampaignCommandHandler(Context);
-            int actual = sut.Handle(new EditCampaignCommand {Campaign = new CampaignSummaryModel { TimeZoneId = "Eastern Standard Time" } });
-            Assert.NotEqual(0, actual);
-        }
-    }
-}
