+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Features/Campaigns/EditCampaignCommand.cs
@@ -3,7 +3,7 @@ using MediatR;
 
 namespace AllReady.Areas.Admin.Features.Campaigns
 {
-    public class EditCampaignCommand : IRequest<int>
+    public class EditCampaignCommand : IAsyncRequest<int>
     {
         public CampaignSummaryModel Campaign {get; set;}
     }
