+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Features/Campaigns/EditCampaignCommandHandler.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using System.Linq;
+using System.Threading.Tasks;
 using AllReady.Areas.Admin.Models;
 using AllReady.Models;
 using MediatR;
@@ -7,23 +7,23 @@ using Microsoft.Data.Entity;
 
 namespace AllReady.Areas.Admin.Features.Campaigns
 {
-    public class EditCampaignCommandHandler : IRequestHandler<EditCampaignCommand, int>
+    public class EditCampaignCommandHandler : IAsyncRequestHandler<EditCampaignCommand, int>
     {
         private AllReadyContext _context;
 
         public EditCampaignCommandHandler(AllReadyContext context)
         {
             _context = context;
-
         }
-        public int Handle(EditCampaignCommand message)
-        {
-            var campaign = _context.Campaigns
-                                    .Include(l => l.Location).ThenInclude(p => p.PostalCode)
-                    .Include(tc => tc.CampaignContacts)
-                    .Include(i => i.CampaignImpact)
 
-                .SingleOrDefault(c => c.Id == message.Campaign.Id);
+        public async Task<int> Handle(EditCampaignCommand message)
+        {
+            var campaign = await _context.Campaigns
+                .Include(l => l.Location).ThenInclude(p => p.PostalCode)
+                .Include(tc => tc.CampaignContacts)
+                .Include(i => i.CampaignImpact)
+                .SingleOrDefaultAsync(c => c.Id == message.Campaign.Id)
+                .ConfigureAwait(false);
 
             if (campaign == null)
             {
@@ -33,31 +33,38 @@ namespace AllReady.Areas.Admin.Features.Campaigns
             campaign.Name = message.Campaign.Name;
             campaign.Description = message.Campaign.Description;
             campaign.FullDescription = message.Campaign.FullDescription;
+
             campaign.TimeZoneId = message.Campaign.TimeZoneId;
-            TimeZoneInfo timeZoneInfo = TimeZoneInfo.FindSystemTimeZoneById(campaign.TimeZoneId);
+
+            var timeZoneInfo = TimeZoneInfo.FindSystemTimeZoneById(campaign.TimeZoneId);
             var startDateTimeOffset = timeZoneInfo.GetUtcOffset(message.Campaign.StartDate);
             campaign.StartDateTime = new DateTimeOffset(message.Campaign.StartDate.Year, message.Campaign.StartDate.Month, message.Campaign.StartDate.Day, 0, 0, 0, startDateTimeOffset);
+
             var endDateTimeOffset = timeZoneInfo.GetUtcOffset(message.Campaign.EndDate);
             campaign.EndDateTime = new DateTimeOffset(message.Campaign.EndDate.Year, message.Campaign.EndDate.Month, message.Campaign.EndDate.Day, 23, 59, 59, endDateTimeOffset);
+
             campaign.ManagingOrganizationId = message.Campaign.OrganizationId;
             campaign.ImageUrl = message.Campaign.ImageUrl;
 
             campaign = campaign.UpdateCampaignContact(message.Campaign, _context);
             campaign.CampaignImpact = campaign.CampaignImpact.UpdateModel(message.Campaign.CampaignImpact);
             campaign.Location = campaign.Location.UpdateModel(message.Campaign.Location);
+
             if (campaign.CampaignImpact != null)
             {
                 _context.Update(campaign.CampaignImpact);
             }
+
             if (campaign.Location != null)
             {
                 _context.Update(campaign.Location);
             }
+
             _context.Update(campaign);
-            _context.SaveChanges();
+
+            await _context.SaveChangesAsync().ConfigureAwait(false);
+
             return campaign.Id;
         }
-
-
     }
 }
